st7789 드라이버 컬러에러 수정

+++ f103_ssd1603/Core/driver/hx711.h
@@ -40,10 +40,10 @@ int32_t     hx711_value(hx711_t *hx711);
 int32_t     hx711_value_ave(hx711_t *hx711, uint16_t sample);
 
 void        hx711_coef_set(hx711_t *hx711, float coef);
-float       hx711_coef_get(hx711_t *hx711);
+int32_t     hx711_coef_get(hx711_t *hx711);
 void        hx711_calibration(hx711_t *hx711, int32_t value_noload, int32_t value_load, float scale);
 void        hx711_tare(hx711_t *hx711, uint16_t sample);
-float       hx711_weight(hx711_t *hx711, uint16_t sample);
+float     hx711_weight(hx711_t *hx711, uint16_t sample);
 void        hx711_power_down(hx711_t *hx711);
 void        hx711_power_up(hx711_t *hx711);
